fix: :lipstick: foregroun notifications adapt for qutoes and admin quotes
