+++ shared/constants/notification-styles.ts
@@ -1,0 +1,22 @@
+export const notificationStyles = {
+  container: {
+    padding: '10px',
+    paddingRight: '15px',
+    paddingLeft: '15px',
+    border: '1px solid #ccc',
+    borderRadius: '5px',
+    backgroundColor: '#f9f9f9',
+  },
+  title: {
+    fontSize: '18px',
+    fontWeight: 'bold',
+    marginBottom: '8px',
+  },
+  item: {
+    marginBottom: '6px',
+  },
+  label: {
+    fontWeight: 'bold',
+    fontSize: '16px',
+  },
+}
